move setup_env to tests module
